correction tasks 6, 8, 9, 11

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -3,7 +3,7 @@ DELIMITER ||
 DROP PROCEDURE IF EXISTS ComputeAverageScoreForUser;
 CREATE PROCEDURE ComputeAverageScoreForUser(IN usr_id INT)
 	BEGIN
-		DECLARE avg_s INT;
+		DECLARE avg_s FLOAT;
 		SELECT AVG(score) FROM corrections WHERE user_id = usr_id INTO avg_s;
 		UPDATE users SET average_score = avg_s WHERE users.id = usr_id;
 	END; ||
